Add rust env var: RUSTC_WRAPPER
Rust's cargo tool can share build artifacts among all packages in a
single workspace, but not across different workspaces. This means that
it is possible for there to be multiple copies of the same build
artifacts from different projects that depend on the same crates.

This change adds the RUSTC_WRAPPER env variable set to the `sccache`
tool. This tool allows sharing build results across different
workspaces.