Conditionally load profiling module
When updating the zsh config with new things being loaded at shell
startup, if startup becomes noticeably slower, it may be hard to
identify what is slowing it down.

This change loads the zsh profiling module based on the value of the
$PROFILE_ZSH environment variable; if the variable equals to the string
"true", then the profiling module is loaded.

Please note that this only profiles zshrc, it does not profile zprofile
or zshenv.